Create locking queues to make sure separate threads can use the queues in parallel safely.
As discussed in https://stackoverflow.com/questions/61310092/vulkan-theaded-application-get-error-message-on-queue-submissions-under-mutex calling device.getQueue(...) is not thread safe, so the changes ensure it is only called once for each queue type. This version is not yet fully thread-safe since the presentQueue of the windows are acquired without interlock. This happens only once, so the concurrent access is not very likely but certainly not impossible.